src/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==22.0.0
idna==3.6
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.41.2
